refactor(webviews): extract tRPC primitives into leaf trpc.ts module
Move trpcToTelemetry, publicProcedureWithTelemetry, the WithTelemetry
helper, and the publicProcedure / router re-exports out of appRouter.ts
into a new leaf module src/webviews/_integration/trpc.ts.

appRouter.ts now imports these primitives from ./trpc instead of
declaring them inline. To keep this commit non-breaking for callers,
appRouter.ts still re-exports them; the follow-up commit migrates each
per-view router to import directly from ./trpc, at which point the
re-exports can be reconsidered.

This is the first half of the fix for the activation-time TDZ error
'Cannot access publicProcedureWithTelemetry before initialization',
introduced by 45f117f when trpc.ts was merged into appRouter.ts. The
cycle is appRouter.ts -> collectionViewRouter.ts -> appRouter.ts: the
per-view router executed at module top-level and accessed
publicProcedureWithTelemetry before appRouter.ts had reached the line
that defines it. Splitting the primitives back out into a leaf module
that nothing else in this folder imports breaks the cycle.
